New Log4Shell logo to raise awareness
There is a seashell made of a scroll (an ancient medium for logging) in the shape of a 'J' to symbolize Java and the 'j' in log4j. The `>` symbolizes the shell prompt that can be accessed by the attacker. The `_` as the cephalopod's pupil symbolizes the shell/terminal cursor. The cephalopod's tentacles spell out `${$$}`, which represents one form of the malicious input string. This image is available under the public domain. https://commons.wikimedia.org/wiki/File:Log4Shell_cephalopod.png